moved types into a types folder, and separated event from user types
