+++ pages/events/nextweek.vue
@@ -3,7 +3,7 @@ import { storeToRefs } from 'pinia';
 import { useEventStore } from '~~/stores/eventStore';
 import useWaitForLogin from '~~/composables/useWaitForLogin';
 import useDayJs from '~~/composables/useDayJs';
-import { WeekEventsInterface, FilteredEventInterface } from '~~/stores/eventTypes';
+import { WeekEventsInterface, FilteredEventInterface } from '~~/types/events';
 
 definePageMeta({
   middleware: ["auth"],
@@ -50,11 +50,11 @@ const filteredEvents: ComputedRef<FilteredEventInterface[]> = computed(() => {
 const config = useRuntimeConfig();
 const sortedEvents = computed(() => {
   const weekEventsHolder: WeekEventsInterface = {
-    Mon: { events:[]},
-    Tue: { events:[]},
-    Wed: { events:[]},
-    Thu: { events:[]},
-    Fri: { events:[]},
+    Mon: { events:[], date: ''},
+    Tue: { events:[], date: ''},
+    Wed: { events:[], date: ''},
+    Thu: { events:[], date: ''},
+    Fri: { events:[], date: ''},
   }
   filteredEvents.value?.filter((event) => event.organizer === config.public.eventOrganizer).forEach((evt) => {
     const day = dayjs(evt.start.dateTime).format('ddd') as 'Mon'|'Tue'|'Wed'|'Thu'|'Fri';
